+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- A new config variable, `bashrc_path`, which is the path to the bashrc script to source.
+
+### Changed
+
+- Removed automatic sourcing of `$HOME/.bashrc` from the SLURM submit script.
+
+### Fixed
+
+- Does not put conda-related lines in SLURM script if `conda_env` is set to `False` or `""`.
+- Changed default config value of `conda_env` from `None` to `""`.
+- A proper `ValueError` will now be raised if `ssh_key_file` is not supplied.
+
 ## [0.15.0] - 2023-05-12
 
 ### Added